hardware_control/requirements.txt
flask
flask-cors
pyserial
picamera2
requests
opencv-python
